Add files via upload
This is a preliminary Gantt Chart I created using Excel. The Column headers are self-explanatory. You can see the Gantt Chart on Sheet 2. It was difficult to display data that does not exist on Excel, therefore I created "Task Examples" just to show what the chart will look like once more entries are made. 

If you would like to change any of the "Task Examples", it should have no degrading effects on the chart, however, if you would like to add more tasks or edit for your own personal use then you will have to manually "Select Data". 

If you want to do this, Fill in the columns accordingly then just follow the steps below. 
1 ) Right Click the Gantt Chart and Select Data
2) Select the Start Date and Press Edit on the "Legend Entries(Series)"
3) Click on the box on top of the "OK" button, you can now extend the data field to include new data.
4) Repeat this for the Duration field.
5) Click "Edit" on the Horizontal Axis and include the new tasks.

Your now entries should now be included in the updated Chart.

I set the Date to be shown in increments of 5 days.

If you want to change this then Right Click the Date on top of the Chart then Format Axis. Under the "Axis Options" change the "Major" to your desired number. 

Again, this is preliminary so feel free to edit/change as you see fit.