Update Node.java
I have finished task 1 from the Scrum Sprint Cycles document and this Node class can now convert the X and Y squared paper format to the screen resolution format. The next task now is to find a way to draw a line between the two nodes. (I have found a link that has the exact answer to this question - https://stackoverflow.com/questions/2902640/android-get-the-screen-resolution-pixels-as-integer-values).